Replace sort 'buttons' with spans
The interactive part of sorting is the entire heading, but by using the 'button' tag for the sort direction arrows indicates to screen readers that the arrows themselves are the interactive content. Not only is this bad practice, but the arrows we are using in Grid.js are below the minimum size requirements for WCAG standard 2.5.8 - 24x24 pixels. By changing the indicators from buttons to spans, we stick to the best practice and avoid the WCAG violation with miniscule changes to the code and no conflicts.